Pair Project
A text adventure, dungeon game in which the player must travel across a generated maze to find the key to the exit. During their journey, they have a chance to battle creatures of both magical and physical as well as obtaining items to assist them. Includes multiple levels, more difficult enemies as levels increase, and an item shop.